single layout for add patient pages
