--- apps/modernization-ui/src/apps/patient/add/basic/add-patient-basic.module.scss
@@ -1,4 +1,0 @@
-.buttonGroup {
-    display: flex;
-    gap: 0.5rem;
-}
