--- apps/modernization-ui/src/apps/patient/add/layout/headerContent/add-patient-header-content.module.scss
@@ -1,14 +1,0 @@
-@use 'styles/borders';
-
-.header {
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-    padding: 1rem;
-    align-items: center;
-    @extend %thin-bottom;
-
-    h1 {
-        margin: 0;
-    }
-}
