+++ apps/modernization-ui/src/apps/patient/add/layout/index.ts
@@ -1,2 +1,1 @@
 export { AddPatientLayout } from './AddPatientLayout';
-export { DataEntryLayout } from './DataEntryLayout';
