+++ apps/modernization-ui/src/design-system/side-nav/side-nav.module.scss
@@ -7,7 +7,6 @@
     height: 100%;
     padding: 1rem;
     gap: 1.5rem;
-    @extend %thin;
 
     .navEntries {
         min-width: 14rem;
