+++ apps/modernization-ui/src/libs/data-entry/DataEntryMenu.spec.tsx
