+++ apps/modernization-ui/src/libs/data-entry/index.ts
@@ -1,0 +1,1 @@
+export { DataEntryLayout } from './DataEntryLayout';
